+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_training.py
@@ -41,7 +41,7 @@ class TestTraining(FormRecognizerTest):
         for doc in model.training_documents:
             self.assertIsNotNone(doc.document_name)
             self.assertIsNotNone(doc.page_count)
-            self.assertEqual(doc.status, "succeeded")
+            self.assertIsNotNone(doc.status)
             self.assertEqual(doc.errors, [])
         for sub in model.submodels:
             self.assertIsNotNone(sub.form_type)
@@ -125,7 +125,7 @@ class TestTraining(FormRecognizerTest):
         for doc in model.training_documents:
             self.assertIsNotNone(doc.document_name)
             self.assertIsNotNone(doc.page_count)
-            self.assertEqual(doc.status, "succeeded")
+            self.assertIsNotNone(doc.status)
             self.assertEqual(doc.errors, [])
         for sub in model.submodels:
             self.assertIsNotNone(sub.form_type)