+++ sdk/search/azure-search-documents/azure/search/documents/_internal/aio/_search_client_async.py
@@ -5,12 +5,10 @@
 # --------------------------------------------------------------------------
 from typing import cast, List, TYPE_CHECKING
 
-import six
-
 from azure.core.tracing.decorator_async import distributed_trace_async
 from ._paging import AsyncSearchItemPaged, AsyncSearchPageIterator
 from .._generated.aio import SearchIndexClient
-from .._generated.models import IndexBatch, IndexingResult, SearchRequest
+from .._generated.models import IndexBatch, IndexingResult
 from .._index_documents_batch import IndexDocumentsBatch
 from .._queries import AutocompleteQuery, SearchQuery, SuggestQuery
 from ..._headers_mixin import HeadersMixin
@@ -104,12 +102,12 @@ class SearchClient(HeadersMixin):
         return cast(dict, result)
 
     @distributed_trace_async
-    async def search(self, query, **kwargs):
-        # type: (Union[str, SearchQuery], **Any) -> AsyncSearchItemPaged[dict]
+    async def search(self, search_text, **kwargs):
+        # type: (str, **Any) -> AsyncSearchItemPaged[dict]
         """Search the Azure search index for documents.
 
-        :param query: An query for searching the index
-        :type documents: str or SearchQuery
+        :param str search_text: A full-text search query expression; Use "*" or omit this parameter to
+        match all documents.
         :rtype:  AsyncSearchItemPaged[dict]
 
         .. admonition:: Example:
@@ -139,14 +137,42 @@ class SearchClient(HeadersMixin):
                 :dedent: 4
                 :caption: Get search result facets.
         """
-        if isinstance(query, six.string_types):
-            query = SearchQuery(search_text=query)
-        elif not isinstance(query, SearchQuery):
-            raise TypeError(
-                "Expected a string or SearchQuery for 'query', but got {}".format(
-                    repr(query)
-                )
-            )
+        include_total_result_count = kwargs.pop("include_total_result_count", None)
+        facets = kwargs.pop("facets", None)
+        filter_arg = kwargs.pop("filter", None)
+        highlight_fields = kwargs.pop("highlight_fields", None)
+        highlight_post_tag = kwargs.pop("highlight_post_tag", None)
+        highlight_pre_tag = kwargs.pop("highlight_pre_tag", None)
+        minimum_coverage = kwargs.pop("minimum_coverage", None)
+        order_by = kwargs.pop("order_by", None)
+        query_type = kwargs.pop("query_type", None)
+        scoring_parameters = kwargs.pop("scoring_parameters", None)
+        scoring_profile = kwargs.pop("scoring_profile", None)
+        search_fields = kwargs.pop("search_fields", None)
+        search_mode = kwargs.pop("search_mode", None)
+        select = kwargs.pop("select", None)
+        skip = kwargs.pop("skip", None)
+        top = kwargs.pop("top", None)
+        query = SearchQuery(
+            search_text=search_text,
+            include_total_result_count=include_total_result_count,
+            facets=facets,
+            filter=filter_arg,
+            highlight_fields=highlight_fields,
+            highlight_post_tag=highlight_post_tag,
+            highlight_pre_tag=highlight_pre_tag,
+            minimum_coverage=minimum_coverage,
+            order_by=order_by,
+            query_type=query_type,
+            scoring_parameters=scoring_parameters,
+            scoring_profile=scoring_profile,
+            search_fields=search_fields,
+            search_mode=search_mode,
+            select=select,
+            skip=skip,
+            top=top
+        )
+
         kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         return AsyncSearchItemPaged(
             self._client, query, kwargs, page_iterator_class=AsyncSearchPageIterator
@@ -154,7 +180,7 @@ class SearchClient(HeadersMixin):
 
     @distributed_trace_async
     async def suggest(self, search_text, suggester_name, **kwargs):
-        # type: (Union[str, SuggestQuery], **Any) -> List[dict]
+        # type: (str, str, **Any) -> List[dict]
         """Get search suggestion results from the Azure search index.
 
         :param str search_text: Required. The search text to use to suggest documents. Must be at least 1
@@ -172,7 +198,28 @@ class SearchClient(HeadersMixin):
                 :dedent: 4
                 :caption: Get search suggestions.
         """
-        query = SuggestQuery(search_text=search_text, suggester_name=suggester_name, **kwargs)
+        filter_arg = kwargs.pop("filter", None)
+        use_fuzzy_matching = kwargs.pop("use_fuzzy_matching", None)
+        highlight_post_tag = kwargs.pop("highlight_post_tag", None)
+        highlight_pre_tag = kwargs.pop("highlight_pre_tag", None)
+        minimum_coverage = kwargs.pop("minimum_coverage", None)
+        order_by = kwargs.pop("order_by", None)
+        search_fields = kwargs.pop("search_fields", None)
+        select = kwargs.pop("select", None)
+        top = kwargs.pop("top", None)
+        query = SuggestQuery(
+            search_text=search_text,
+            suggester_name=suggester_name,
+            filter=filter_arg,
+            use_fuzzy_matching=use_fuzzy_matching,
+            highlight_post_tag=highlight_post_tag,
+            highlight_pre_tag=highlight_pre_tag,
+            minimum_coverage=minimum_coverage,
+            order_by=order_by,
+            search_fields=search_fields,
+            select=select,
+            top=top
+        )
 
         kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         response = await self._client.documents.suggest_post(
@@ -200,7 +247,26 @@ class SearchClient(HeadersMixin):
                 :dedent: 4
                 :caption: Get a auto-completions.
         """
-        query = AutocompleteQuery(search_text=search_text, suggester_name=suggester_name, **kwargs)
+        autocomplete_mode = kwargs.pop("autocomplete_mode", None)
+        filter_arg = kwargs.pop("filter", None)
+        use_fuzzy_matching = kwargs.pop("use_fuzzy_matching", None)
+        highlight_post_tag = kwargs.pop("highlight_post_tag", None)
+        highlight_pre_tag = kwargs.pop("highlight_pre_tag", None)
+        minimum_coverage = kwargs.pop("minimum_coverage", None)
+        search_fields = kwargs.pop("search_fields", None)
+        top = kwargs.pop("top", None)
+        query = AutocompleteQuery(
+            search_text=search_text,
+            suggester_name=suggester_name,
+            autocomplete_mode=autocomplete_mode,
+            filter=filter_arg,
+            use_fuzzy_matching=use_fuzzy_matching,
+            highlight_post_tag=highlight_post_tag,
+            highlight_pre_tag=highlight_pre_tag,
+            minimum_coverage=minimum_coverage,
+            search_fields=search_fields,
+            top=top
+        )
 
         kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         response = await self._client.documents.autocomplete_post(
@@ -231,7 +297,7 @@ class SearchClient(HeadersMixin):
                 :caption: Upload new documents to an index
         """
         batch = IndexDocumentsBatch()
-        batch.add_upload_documents(documents)
+        batch.add_upload_actions(documents)
 
         kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         results = await self.index_documents(batch, **kwargs)
@@ -264,7 +330,7 @@ class SearchClient(HeadersMixin):
                 :caption: Delete existing documents to an index
         """
         batch = IndexDocumentsBatch()
-        batch.add_delete_documents(documents)
+        batch.add_delete_actions(documents)
 
         kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         results = await self.index_documents(batch, **kwargs)
@@ -293,7 +359,7 @@ class SearchClient(HeadersMixin):
                 :caption: Merge fields into existing documents to an index
         """
         batch = IndexDocumentsBatch()
-        batch.add_merge_documents(documents)
+        batch.add_merge_actions(documents)
 
         kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         results = await self.index_documents(batch, **kwargs)
@@ -313,7 +379,7 @@ class SearchClient(HeadersMixin):
         :rtype:  List[IndexingResult]
         """
         batch = IndexDocumentsBatch()
-        batch.add_merge_or_upload_documents(documents)
+        batch.add_merge_or_upload_actions(documents)
 
         kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         results = await self.index_documents(batch, **kwargs)