+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.0.1 (Unreleased)
+
+- Fixed error raised when a non string ID is used in an item. It now raises TypeError rather than AttributeError. Issue 11793 - thank you @Rabbit994.
+
+
 ## 4.0.0 (2020-05-20)
 
 - Stable release.