+++ sdk/servicebus/azure-servicebus/CHANGELOG.md
@@ -1,9 +1,11 @@
 # Release History
 
-## 7.0.0b3 (Unreleased)
+## 7.0.0b3 (2020-06-08)
 
 **New Features**
 
+* Added support for management of queue entities.
+    - Use `azure.servicebus.management.ServiceBusManagementClient` (`azure.servicebus.management.aio.ServiceBusManagementClient` for aio) to create, update, delete, list queues and get settings as well as runtime information of queues under a ServiceBus namespace.
 * Added methods `get_queue_deadletter_receiver` and `get_subscription_deadletter_receiver` in `ServiceBusClient` to get a `ServiceBusReceiver` for the dead-letter sub-queue of the target entity.
 
 **BugFixes**